Refactored and Cleaned Up Dashboard , PlanningTable A, ResultsTabe A
